Cmake fixes

+++ include/cdcpd/past_template_matcher.h
@@ -11,14 +11,14 @@
 
 class PastTemplateMatcher {
 public:
-    PastTemplateMatcher(int _sample_size);
+    PastTemplateMatcher(size_t _sample_size);
 
     size_t size();
 
     // TODO returning vectors of big matrices is very inefficient. Perhaps return pointers, or something, instead.
     std::vector<Eigen::Matrix3Xf> query_template(pcl::PointCloud<pcl::PointXYZ>::ConstPtr filtered_points, int k=8);
 
-    void add_template(pcl::PointCloud<pcl::PointXYZ>::Ptr filtered_points, Eigen::Matrix3Xf tracking_result); 
+    void add_template(pcl::PointCloud<pcl::PointXYZ>::Ptr filtered_points, Eigen::Matrix3Xf tracking_result);
 
 private:
     cv::Mat downsampled_feature(pcl::PointCloud<pcl::PointXYZ>::ConstPtr filtered_points);
@@ -30,5 +30,6 @@ private:
     std::random_device rd;
     std::mt19937 gen;
     std::uniform_int_distribution<> index_generator;
-    int sample_size;
+    size_t sample_size;
 };
+
